requirements.txt
PyYAML==6.0
urllib3==1.26.12
